Updated position so RPM returns a float

+++ ECU2018/ECU2018/position.cpp
@@ -66,11 +66,11 @@ int encoderPositionEngine()
 	return encoder_A - calibration_variable; // correcting to TDC2 is 0
 }
 
-int encoderRPM() { // Returns the RPM. Returns -1 every ~70min
+float encoderRPM() { // Returns the RPM. Returns -1 every ~70min
 	if (encoder_Z_time - encoder_Z_time_old <= 0) {
 		return -1;
 	}
-	int RPM = (int)(60000000 / (float)(encoder_Z_time - encoder_Z_time_old));
+	float RPM = (60000000 / (float)(encoder_Z_time - encoder_Z_time_old));
 	// T = encoder_Z_time - encoder_Z_time_old // How many µs between 2 ticks
 	// 60 s/min * 10e6 µs/s  /  T µs   =  60000000 / T  rounds/min
 	return RPM;
